# backend-api-spec.yaml
openapi: 3.0.3
info:
  title: CityPulse Energy Management API
  description: Comprehensive API for CityPulse Energy Management System
  version: 1.0.0
  contact:
    name: Marc (Backend Developer)
    email: marc@citypulse.com
servers:
  - url: https://api.citypulse.com/v1
    description: Production server
  - url: http://localhost:8000/v1
    description: Development server
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
...
  /energy/production:
    get:
      tags:
        - Energy Data
      summary: Get energy production data (solar)
      parameters:
        - name: building_id
          in: query
          schema:
            type: string
        - name: start_date
          in: query
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          schema:
            type: string
            format: date-time
        - name: granularity
          in: query
          schema:
            type: string
            enum: [hour, day, week, month]
            default: hour
      responses:
        '200':
          description: Energy production data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EnergyData'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
        user:
          $ref: '#/components/schemas/User'

    # User Schemas
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [admin, manager, user]
        avatar:
          type: string
          format: url
        created_at:
          type: string
          format: date-time
        last_login:
          type: string
          format: date-time
...
    EnergyData:
      type: object
      properties:
        id:
          type: string
        building_id:
          type: string
        timestamp:
          type: string
          format: date-time
        consumption:
          type: number
          description: kWh consumed
        production:
          type: number
          description: kWh produced (solar)
        efficiency:
          type: number
          description: Percentage efficiency
        co2_saved:
          type: number
          description: kg CO2 saved

    # Analytics Schemas
...